+++ packages/terra-date-picker/tests/jest/DatePicker.test.jsx
@@ -50,3 +50,17 @@ it('should render a date picker with min and max dates', () => {
   const wrapper = shallow(datePicker);
   expect(wrapper).toMatchSnapshot();
 });
+
+it('should render a date picker with releaseFocus', () => {
+  const releaseFocus = () => {};
+  const datePicker = (<DatePicker name="date-input" releaseFocus={releaseFocus} utcOffset={0} />);
+  const wrapper = shallow(datePicker);
+  expect(wrapper).toMatchSnapshot();
+});
+
+it('should render a date picker with requestFocus', () => {
+  const requestFocus = () => {};
+  const datePicker = (<DatePicker name="date-input" requestFocus={requestFocus} utcOffset={0} />);
+  const wrapper = shallow(datePicker);
+  expect(wrapper).toMatchSnapshot();
+});