+++ packages/terra-site/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added new example to display date picker and popup inside modal.
 
 1.2.0 - (July 26, 2017)
 ------------------